Road interconnection and filtering

+++ memory_obj.h
@@ -46,8 +46,8 @@ void mem_CloseLog();
 short *mem_GetGRCrowptr ();
 short *mem_GetGRCcolptr ();
 
-short *mem_GetRERCrowptr();
-short *mem_GetRERCcolptr();
+int *mem_GetRERCrowptr();
+int *mem_GetRERCcolptr();
 
 /* D.D. Added for cumulative growth array - 8/17/2006                */
 short *mem_GetGRZrowptr ();
